Namespace policies for manage pages Fixes #222
