--- app/policies/active_admin/page_policy.rb
@@ -1,12 +1,0 @@
-module ActiveAdmin
-  class PagePolicy < ApplicationPolicy
-    def show?
-      case record.name
-      when "Dashboard"
-        true
-      else
-        false
-      end
-    end
-  end
-end
