+++ app/policies/discharge_policy.rb
@@ -1,30 +1,9 @@
-class DischargePolicy < ApplicationPolicy
+class DischargePolicy < Manage::DischargePolicy
   def index?
     true
   end
 
   def show?
     true
-  end
-
-  def new?
-    user&.has_permission?("discharge_add") ||
-      user&.has_permission?("discharge_add_any")
-  end
-
-  def create?
-    user&.has_permission_on_user?("discharge_add", record.user) ||
-      user&.has_permission?("discharge_add_any") ||
-      user&.has_permission?("admin")
-  end
-
-  def update?
-    user&.has_permission_on_user?("discharge_add", record.user) ||
-      user&.has_permission?("discharge_add_any") ||
-      user&.has_permission?("admin")
-  end
-
-  def destroy?
-    user&.has_permission?("admin")
   end
 end
