+++ app/policies/event_policy.rb
@@ -1,30 +1,10 @@
-class EventPolicy < ApplicationPolicy
+class EventPolicy < Manage::EventPolicy
   def index?
     user&.member? || user&.honorably_discharged?
   end
 
   def show?
     user&.member?
-  end
-
-  def new?
-    user&.has_permission?("event_add") ||
-      user&.has_permission?("event_add_any") ||
-      user&.has_permission?("admin")
-  end
-
-  def create?
-    (record.unit && user&.has_permission_on_unit?("event_add", record.unit)) ||
-      user&.has_permission?("event_add_any") ||
-      user&.has_permission?("admin")
-  end
-
-  def update?
-    create?
-  end
-
-  def destroy?
-    create?
   end
 
   def aar?
