+++ app/policies/finance_record_policy.rb
@@ -1,26 +1,9 @@
-class FinanceRecordPolicy < ApplicationPolicy
+class FinanceRecordPolicy < Manage::FinanceRecordPolicy
   def index?
     user&.member?
   end
 
   def show?
     index?
-  end
-
-  def new?
-    user&.has_permission?("finance_add") ||
-      user&.has_permission?("admin")
-  end
-
-  def create?
-    new?
-  end
-
-  def update?
-    new?
-  end
-
-  def destroy?
-    new?
   end
 end
