+++ app/policies/manage/active_admin/page_policy.rb
@@ -1,0 +1,14 @@
+module Manage
+  module ActiveAdmin
+    class PagePolicy < ApplicationPolicy
+      def show?
+        case record.name
+        when "Dashboard"
+          true
+        else
+          false
+        end
+      end
+    end
+  end
+end
