+++ app/policies/manage/award_policy.rb
@@ -1,10 +1,10 @@
-class RestrictedNamePolicy < ApplicationPolicy
+class Manage::AwardPolicy < ApplicationPolicy
   def index?
-    user&.has_permission?("admin")
+    create?
   end
 
   def show?
-    user&.has_permission?("admin")
+    create?
   end
 
   def create?
