+++ app/policies/manage/ban_log_policy.rb
@@ -1,0 +1,26 @@
+class Manage::BanLogPolicy < ApplicationPolicy
+  def index?
+    new?
+  end
+
+  def show?
+    new?
+  end
+
+  def new?
+    user&.has_permission?("banlog_edit_any") ||
+      user&.has_permission?("admin")
+  end
+
+  def create?
+    new?
+  end
+
+  def update?
+    new?
+  end
+
+  def destroy?
+    user&.has_permission?("admin")
+  end
+end
