+++ app/policies/manage/permission_policy.rb
@@ -1,0 +1,21 @@
+class Manage::PermissionPolicy < ApplicationPolicy
+  def index?
+    user&.has_permission?("admin")
+  end
+
+  def show?
+    user&.has_permission?("admin")
+  end
+
+  def create?
+    user&.has_permission?("admin")
+  end
+
+  def update?
+    user&.has_permission?("admin")
+  end
+
+  def destroy?
+    user&.has_permission?("admin")
+  end
+end
