+++ app/policies/manage/rank_policy.rb
@@ -1,0 +1,21 @@
+class Manage::RankPolicy < ApplicationPolicy
+  def index?
+    create?
+  end
+
+  def show?
+    create?
+  end
+
+  def create?
+    user&.has_permission?("admin")
+  end
+
+  def update?
+    user&.has_permission?("admin")
+  end
+
+  def destroy?
+    user&.has_permission?("admin")
+  end
+end
