+++ app/policies/position_policy.rb
@@ -1,21 +1,9 @@
-class PositionPolicy < ApplicationPolicy
+class PositionPolicy < Manage::PositionPolicy
   def index?
     true
   end
 
   def show?
     true
-  end
-
-  def create?
-    user&.has_permission?("admin")
-  end
-
-  def update?
-    user&.has_permission?("admin")
-  end
-
-  def destroy?
-    user&.has_permission?("admin")
   end
 end
